requirements.txt
python=3.9.8
Django=4.0.1
pillow=9.0.0
psycopg2=2.9.3
django-bootstrap-v5
python-binance=1.0.15

